🔥 del : Body.tsx Cell.tsx 변경 및 코드 리팩토링함
