--- src/components/sales/status/tab/calendar/body/Body.tsx
@@ -1,5 +1,0 @@
-const Body = () => {
-  return <div></div>;
-};
-
-export default Body;
